+++ CHANGELOG.markdown
@@ -1,6 +1,7 @@
 0.7.0.0
 -------
 * Fix date conversion. `recordedEventCreateEpoch` is no longer exposed.
+* Add `waitTillCatchup` and `hasCaughtUp` functions.
 
 0.6.0.1
 -------